Unstable commit during rework on reward and block calculation converting from double to bigdecimal for more precision

+++ adrestus-core/src/main/java/io/Adrestus/core/RegularTransaction.java
@@ -2,6 +2,7 @@ package io.Adrestus.core;
 
 import io.Adrestus.crypto.elliptic.ECDSASignatureData;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 
 public class RegularTransaction extends Transaction {
@@ -10,15 +11,15 @@ public class RegularTransaction extends Transaction {
         super(TransactionType.REGULAR);
     }
 
-    public RegularTransaction(String hash, TransactionType type, StatusType status, int zoneFrom, int zoneTo, String timestamp, int blockNumber, String from, String to, double amount, double AmountWithTransactionFee, int nonce, ECDSASignatureData signature) {
+    public RegularTransaction(String hash, TransactionType type, StatusType status, int zoneFrom, int zoneTo, String timestamp, int blockNumber, String from, String to, BigDecimal amount, BigDecimal AmountWithTransactionFee, int nonce, ECDSASignatureData signature) {
         super(hash, type, status, zoneFrom, zoneTo, timestamp, blockNumber, from, to, amount, AmountWithTransactionFee, nonce, signature);
     }
 
-    public RegularTransaction(String hash, TransactionType type, StatusType status, int zoneFrom, int zoneTo, String timestamp, String from, String to, double amount, double AmountWithTransactionFee, int nonce, ECDSASignatureData signature) {
+    public RegularTransaction(String hash, TransactionType type, StatusType status, int zoneFrom, int zoneTo, String timestamp, String from, String to, BigDecimal amount, BigDecimal AmountWithTransactionFee, int nonce, ECDSASignatureData signature) {
         super(hash, type, status, zoneFrom, zoneTo, timestamp, from, to, amount, AmountWithTransactionFee, nonce, signature);
     }
 
-    public RegularTransaction(String hash, TransactionType type, StatusType status, int zoneFrom, int zoneTo, String timestamp, int blockNumber, String from, String to, double amount, double AmountWithTransactionFee, int nonce, BigInteger XAxis, BigInteger YAxis, ECDSASignatureData signature) {
+    public RegularTransaction(String hash, TransactionType type, StatusType status, int zoneFrom, int zoneTo, String timestamp, int blockNumber, String from, String to, BigDecimal amount, BigDecimal AmountWithTransactionFee, int nonce, BigInteger XAxis, BigInteger YAxis, ECDSASignatureData signature) {
         super(hash, type, status, zoneFrom, zoneTo, timestamp, blockNumber, from, to, amount, AmountWithTransactionFee, nonce, XAxis, YAxis, signature);
     }
