+++ adrestus-core/src/main/java/io/Adrestus/core/RewardsTransaction.java
@@ -3,6 +3,7 @@ package io.Adrestus.core;
 import io.Adrestus.crypto.elliptic.ECDSASignatureData;
 import io.activej.serializer.annotations.Serialize;
 
+import java.math.BigDecimal;
 import java.util.Objects;
 
 
@@ -18,12 +19,12 @@ public class RewardsTransaction extends Transaction {
         this.RecipientAddress = RecipientAddress;
     }
 
-    public RewardsTransaction(String hash, TransactionType type, StatusType status, int zoneFrom, int zoneTo, String timestamp, int blockNumber, String from, String to, double amount, double transactionFee, int nonce, ECDSASignatureData signature, String RecipientAddress) {
+    public RewardsTransaction(String hash, TransactionType type, StatusType status, int zoneFrom, int zoneTo, String timestamp, int blockNumber, String from, String to, BigDecimal amount, BigDecimal transactionFee, int nonce, ECDSASignatureData signature, String RecipientAddress) {
         super(hash, type, status, zoneFrom, zoneTo, timestamp, blockNumber, from, to, amount, transactionFee, nonce, signature);
         this.RecipientAddress = RecipientAddress;
     }
 
-    public RewardsTransaction(TransactionType type, String RecipientAddress, double amount) {
+    public RewardsTransaction(TransactionType type, String RecipientAddress, BigDecimal amount) {
         super(type, amount);
         this.RecipientAddress = RecipientAddress;
     }
