+++ adrestus-network/src/test/java/io/Adrestus/p2p/kademlia/RandomIPTest.java
@@ -35,6 +35,7 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.spongycastle.util.encoders.Hex;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.security.SecureRandom;
 import java.util.ArrayList;
@@ -74,7 +75,7 @@ public class RandomIPTest {
 
         kademliaData = new KademliaData(new SecurityAuditProofs(adddress, ecKeyPair.getPublicKey(), signatureData));
         kademliaData.getAddressData().setValidatorBlSPublicKey(vk);
-        TreeFactory.getMemoryTree(0).store(adddress, new PatriciaTreeNode(1000, 0));
+        TreeFactory.getMemoryTree(0).store(adddress, new PatriciaTreeNode(BigDecimal.valueOf(1000), 0));
         Gson gson = new Gson();
         String jsonString = gson.toJson(kademliaData);
         KademliaData copydata = gson.fromJson(jsonString, KademliaData.class);
